properly divide implementation
